warnings--
svn path=/trunk/extragear/multimedia/amarok/; revision=695975

+++ src/context/Svg.cpp
@@ -27,12 +27,14 @@ Svg::Svg( const QString& imagePath, QObject* parent )
 QRect Svg::elementRect( const QString& elementId )
 {
     QRect rect = Plasma::Svg::elementRect( elementId );
-    debug() << "original rect of element: " << elementId << " " << rect << " transform matrix: " << matrixForElement( elementId ) << " new rect: " << rect.translated( matrixForElement( elementId ).dx(), matrixForElement( elementId ).dy() );
-    rect.translate( matrixForElement( elementId ).dx(), matrixForElement( elementId ).dy() );
-
+    debug() << "original rect of element: " << elementId << " " << rect 
+            << " transform matrix: " << matrixForElement( elementId ) 
+            << " new rect: " << rect.translated( (int)matrixForElement( elementId ).dx(), (int)matrixForElement( elementId ).dy() );
+    rect.translate( (int)matrixForElement( elementId ).dx(), (int)matrixForElement( elementId ).dy() );
 
     return rect;
 
 }
 
 } // namespace Context
+
